feat: Email VO 객체 도입으로 도메인 책임 분리
- Email VO 클래스 생성하여 이메일 검증 로직 분리
- User 엔티티에서 Email VO 활용으로 책임 명확화
- User 엔티티의 validate 메서드 통합 및 중복 제거
