+++ apps/commerce-api/src/main/java/com/loopers/domain/example/User.java
@@ -18,8 +18,8 @@ public class User extends BaseEntity {
     @Id
     private String userId;
 
-    @Column(nullable = false)
-    private String email;
+    @Embedded
+    private Email email;
 
     @Column(nullable = false)
     private LocalDate birthday;
@@ -28,39 +28,28 @@ public class User extends BaseEntity {
     @Enumerated(EnumType.STRING)
     private Gender gender;
 
-    @Column(nullable = false)
-    private int point;
 
 
     public User(String userId, String email, String birthday, Gender gender) {
-        validateUserId(userId);
-        validateEmail(email);
-        validateBirthday(birthday);
-        validateGender(gender);
+        validate(userId, email, birthday, gender);
 
         this.userId = userId;
-        this.email = email;
+        this.email = new Email(email);
         this.birthday = parseBirthday(birthday);
         this.gender = gender;
     }
 
-    public User(String userId, String email, String birthday, Gender gender, int point) {
-        validateUserId(userId);
-        validateEmail(email);
-        validateBirthday(birthday);
-        validateGender(gender);
-        validatePoint(point);
-
-        this.userId = userId;
-        this.email = email;
-        this.birthday = parseBirthday(birthday);
-        this.gender = gender;
-        this.point = point;
-    }
 
 
 
     protected User() {}
+
+    private void validate(String userId, String email, String birthday, Gender gender) {
+        validateUserId(userId);
+        // Email 검증은 Email VO 생성자에서 처리
+        validateBirthday(birthday);
+        validateGender(gender);
+    }
 
     private void validateUserId(String userId) {
         if (userId == null || userId.isBlank()) {
@@ -74,14 +63,6 @@ public class User extends BaseEntity {
         }
     }
 
-    private void validateEmail(String email) {
-        if (email == null || email.isBlank()) {
-            throw new CoreException(ErrorType.INVALID_EMAIL);
-        }
-        if (!isValidEmailFormat(email)) {
-            throw new CoreException(ErrorType.INVALID_EMAIL_FORMAT);
-        }
-    }
 
     private void validateBirthday(String birthday) {
         if (birthday == null || birthday.isBlank()) {
@@ -98,24 +79,8 @@ public class User extends BaseEntity {
         return LocalDate.parse(birthday, DateTimeFormatter.ofPattern("yyyy-MM-dd"));
     }
 
-    private boolean isValidEmailFormat(String email) {
-        int atIndex = email.indexOf('@');
-        if (atIndex <= 0 || atIndex >= email.length() - 1) {
-            return false;
-        }
-
-        String domain = email.substring(atIndex + 1);
-        int dotIndex = domain.lastIndexOf('.');
-
-        return dotIndex > 0 && dotIndex < domain.length() - 1;
-    }
     private void validateGender(Gender gender) {
         if (gender == null) {
-            throw new CoreException(ErrorType.INVALID_GENDER);
-        }
-    }
-    private void validatePoint(int point) {
-        if (point < 0){
             throw new CoreException(ErrorType.INVALID_GENDER);
         }
     }
@@ -125,7 +90,7 @@ public class User extends BaseEntity {
     }
 
     public String getEmail() {
-        return email;
+        return email.getValue();
     }
 
     public LocalDate getBirthday(){
@@ -133,12 +98,5 @@ public class User extends BaseEntity {
     }
     public Gender getGender() {
         return gender;
-    }
-    public int getPoint(){
-        return point;
-    }
-    public void setPoint(int point) {
-        validatePoint(point);  // 유효성 검증
-        this.point = point;    // 값 설정
     }
 }
